+++ src/composable/index.ts
@@ -3,4 +3,4 @@ export { useTaskOperationMessage } from './useTaskOperationMessage'
 export { useTaskRightContextMenu } from './useTaskRightContextMenu'
 export { useTaskSidebarDrag } from './useTaskSidebarDrag'
 export { isDark, naiveTheme, toggleDark } from './useTheme'
-export { useTaskLeftListCreateProject } from './useTaskLeftListAddProject'
+export { useTaskLeftListCreateProject } from './useTaskLeftListCreateProject'
